Melhoria no tratamento de erros 400

+++ src/main/java/br/com/alura/livraria/service/AutorService.java
@@ -4,6 +4,7 @@ import javax.persistence.EntityNotFoundException;
 
 import org.modelmapper.ModelMapper;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
@@ -14,12 +15,16 @@ import br.com.alura.livraria.dto.AutorFormDto;
 import br.com.alura.livraria.dto.AutorOutputDto;
 import br.com.alura.livraria.modelo.Autor;
 import br.com.alura.livraria.repository.AutorRepository;
+import br.com.alura.livraria.repository.LivroRepository;
 
 @Service
 public class AutorService {
 	
 	@Autowired
 	private AutorRepository autorRepository;
+	
+	@Autowired
+	private LivroRepository livroRepository;
 	
 	@Autowired
 	private ModelMapper modelMapper;
@@ -31,6 +36,11 @@ public class AutorService {
 
 	@Transactional
 	public AutorOutputDto cadastrar(AutorFormDto autorFormDto) {
+		boolean temEmailCadastrado = autorRepository.existsByEmail(autorFormDto.getEmail());
+		if(temEmailCadastrado) {
+			throw new DataIntegrityViolationException("E-mail indisponível!");
+		}
+		
 		Autor autor = modelMapper.map(autorFormDto, Autor.class);
 		
 		autorRepository.save(autor);
@@ -56,6 +66,10 @@ public class AutorService {
 	
 	@Transactional
 	public void remover(Long id) {
+		boolean temLivrosCadastrados = livroRepository.existsByAutorId(id);
+		if(temLivrosCadastrados) {
+			throw new DataIntegrityViolationException("Autor possuí Livros registrados. Não pode ser excluído!");
+		}
 		autorRepository.deleteById(id);
 	}
